Создал программку по ДЗ (Задача №1), закомитил

+++ Lesson1/HomeWork/Program.cs
@@ -1,6 +1,6 @@
 ﻿int a = 5;
 int b = 7;
 int c = 9;
-int d = 11;
-long resultat = a*b/(c+d);
+float d = 11;
+float resultat = a*b/(c+d);
 Console.WriteLine(resultat);
